+++ backend/src/routes/alerts.ts
@@ -1,0 +1,55 @@
+import { Router, type Request, type Response } from 'express';
+import { z } from 'zod';
+import { DatabaseService } from '../services/DatabaseService';
+import { authMiddleware } from '../middleware/auth';
+import { requireAdmin } from '../middleware/tierGates';
+
+const AlertCreateSchema = z.object({
+  stack_name: z.string().min(1).max(255),
+  metric: z.enum(['cpu_percent', 'memory_percent', 'memory_mb', 'net_rx', 'net_tx', 'restart_count']),
+  operator: z.enum(['>', '>=', '<', '<=', '==']),
+  threshold: z.number().min(0),
+  duration_mins: z.coerce.number().int().min(0).max(1440),
+  cooldown_mins: z.coerce.number().int().min(0).max(10080),
+});
+
+export const alertsRouter = Router();
+
+alertsRouter.get('/', authMiddleware, async (req: Request, res: Response) => {
+  try {
+    let stackName = req.query.stackName as string | undefined;
+    if (Array.isArray(stackName)) stackName = stackName[0] as string;
+
+    const alerts = DatabaseService.getInstance().getStackAlerts(stackName);
+    res.json(alerts);
+  } catch {
+    res.status(500).json({ error: 'Failed to fetch alerts' });
+  }
+});
+
+alertsRouter.post('/', authMiddleware, async (req: Request, res: Response) => {
+  if (!requireAdmin(req, res)) return;
+  const parsed = AlertCreateSchema.safeParse(req.body);
+  if (!parsed.success) {
+    res.status(400).json({ error: 'Invalid alert data', details: parsed.error.flatten().fieldErrors });
+    return;
+  }
+  try {
+    const created = DatabaseService.getInstance().addStackAlert(parsed.data);
+    res.status(201).json(created);
+  } catch (error) {
+    console.error('Failed to add alert:', error);
+    res.status(500).json({ error: 'Failed to add alert' });
+  }
+});
+
+alertsRouter.delete('/:id', authMiddleware, async (req: Request, res: Response) => {
+  if (!requireAdmin(req, res)) return;
+  try {
+    const id = parseInt(req.params.id as string, 10);
+    DatabaseService.getInstance().deleteStackAlert(id);
+    res.json({ success: true });
+  } catch {
+    res.status(500).json({ error: 'Failed to delete alert' });
+  }
+});
